+++ src/components/FilterButton/index.ts
@@ -1,0 +1,1 @@
+export { default as FilterButton } from './FilterButton';
