--- src/components/SportsButton/SportsButton.style.ts
@@ -1,9 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-import styled from '@emotion/styled';
-
-export const Container = styled.button`
-  position: relative;
-  display: inline-block;
-  border: none;
-  border-radius: 50%;
-`;
