--- src/components/SportsButton/index.ts
@@ -1,2 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-export { default as SportsButton } from './SportsButton';
